+++ source/database/models/companies/Company.js
@@ -41,14 +41,6 @@ class Company extends Model {
 		return Math.floor(Math.sqrt(this.getXP(hunters) / 3) + 1);
 	}
 
-	getThumbnailURLMap() {
-		return {
-			open: this.openBountyThumbnailURL ?? "https://cdn.discordapp.com/attachments/545684759276421120/734093574031016006/bountyboard.png",
-			complete: this.completedBountyThumbnailURL ?? "https://cdn.discordapp.com/attachments/545684759276421120/734092918369026108/completion.png",
-			deleted: this.deletedBountyThumbnailURL ?? "https://cdn.discordapp.com/attachments/545684759276421120/734093574031016006/bountyboard.png"
-		};
-	}
-
 	festivalMultiplierString() {
 		if (this.festivalMultiplier != 1) {
 			return ` ${bold(italic(`x${this.festivalMultiplier}`))}`;
@@ -113,16 +105,34 @@ function initModel(sequelize) {
 			type: DataTypes.STRING
 		},
 		openBountyThumbnailURL: {
-			type: DataTypes.STRING
+			type: DataTypes.STRING,
+			get() {
+				return this.openBountyThumbnailURL ?? "https://cdn.discordapp.com/attachments/545684759276421120/734093574031016006/bountyboard.png";
+			}
 		},
 		completedBountyThumbnailURL: {
-			type: DataTypes.STRING
+			type: DataTypes.STRING,
+			get() {
+				return this.completedBountyThumbnailURL ?? "https://cdn.discordapp.com/attachments/545684759276421120/734092918369026108/completion.png";
+			}
 		},
 		deletedBountyThumbnailURL: {
-			type: DataTypes.STRING
+			type: DataTypes.STRING,
+			get() {
+				return this.deletedBountyThumbnailURL ?? "https://cdn.discordapp.com/attachments/545684759276421120/734093574031016006/bountyboard.png";
+			}
 		},
 		scoreboardThumbnailURL: {
-			type: DataTypes.STRING
+			type: DataTypes.STRING,
+			get() {
+				return this.scoreboardThumbnailURL ?? "https://cdn.discordapp.com/attachments/545684759276421120/734094693217992804/scoreboard.png";
+			}
+		},
+		goalCompletionThumbnailURL: {
+			type: DataTypes.STRING,
+			get() {
+				return this.goalCompletionThumbnailURL ?? "https://cdn.discordapp.com/attachments/673600843630510123/1309260766318166117/trophy-cup.png?ex=6740ef9b&is=673f9e1b&hm=218e19ede07dcf85a75ecfb3dde26f28adfe96eb7b91e89de11b650f5c598966&";
+			}
 		}
 	}, {
 		sequelize,